+++ metadata.rb
@@ -1,10 +1,10 @@
 name             'ca_openldap'
 maintainer       'Christophe Arguel'
 maintainer_email 'christophe.arguel@free.fr'
-license          'All rights reserved'
-description      'Installs/Configures Openldap (>=2.4)'
+license          'Apache 2.0'
+description      'Configures a node to be an OpenLDAP server or client.'
 long_description IO.read(File.join(File.dirname(__FILE__), 'README.md'))
-version          '0.2.1'
+version          '0.2.2'
 
 supports 'redhat', '>= 6.0'
 supports 'centos', '>= 6.0'