+++ recipes/server.rb
@@ -57,12 +57,20 @@ end
 
 # TLS connection configuration
 (use_ldap, use_ldaps) = use_ldap_or_ldaps?(node.ca_openldap.tls.enable)
+ldap_port = node.ca_openldap.default_ports.ldap
+ldaps_port = node.ca_openldap.default_ports.ldaps
+
+urls = []
+urls << "ldap://*:#{ldap_port}" if use_ldap == "yes"
+urls << "ldaps://*:#{ldaps_port}" if use_ldaps == "yes"
 
 ruby_block "tls_connection_configuration" do
   block do
     f = Chef::Util::FileEdit.new("/etc/sysconfig/ldap")
-    f.search_file_replace_line(/SLAPD_LDAP=/, "SLAPD_LDAP=#{use_ldap}")
-    f.search_file_replace_line(/SLAPD_LDAPS=/, "SLAPD_LDAPS=#{use_ldaps}")
+    f.search_file_replace_line(/SLAPD_LDAP=/, "SLAPD_LDAP=no")
+    f.search_file_replace_line(/SLAPD_LDAPS=/, "SLAPD_LDAPS=no")
+    f.search_file_replace_line(/SLAPD_LDAPI=/, "SLAPD_LDAPI=yes")
+    f.search_file_replace_line(/SLAPD_URLS=/, "SLAPD_URLS=\"#{urls.join ""}\"")
     f.write_file
   end
 end